[core] phpunit

+++ src/Controller/CategoryController.php
@@ -10,6 +10,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
+
 /**
  * @Route("/category")
  */
@@ -20,8 +21,8 @@ class CategoryController extends AbstractController
      */
     public function index(CategoryRepository $categoryRepository): Response
     {
-
-
+        $hasAccess = $this->isGranted('ROLE_USER');
+        $this->denyAccessUnlessGranted('ROLE_USER');
 
         return $this->render('category/index.html.twig', [
             'categories' => $categoryRepository->findAll(),
@@ -33,7 +34,8 @@ class CategoryController extends AbstractController
      */
     public function new(Request $request): Response
     {
-
+        $hasAccess = $this->isGranted('ROLE_ADMIN');
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
@@ -58,6 +60,9 @@ class CategoryController extends AbstractController
      */
     public function show(Category $category): Response
     {
+        $hasAccess = $this->isGranted('ROLE_ADMIN');
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         return $this->render('category/show.html.twig', [
             'category' => $category,
         ]);
@@ -68,6 +73,9 @@ class CategoryController extends AbstractController
      */
     public function edit(Request $request, Category $category): Response
     {
+        $hasAccess = $this->isGranted('ROLE_ADMIN');
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
 
@@ -88,6 +96,9 @@ class CategoryController extends AbstractController
      */
     public function delete(Request $request, Category $category): Response
     {
+        $hasAccess = $this->isGranted('ROLE_SUPERADMIN');
+        $this->denyAccessUnlessGranted('ROLE_SUPERADMIN');
+
         if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($category);
